Split out components so it's easier to read the code.
